Rename linear regression files
