+++ lin_reg.h
